Switch to testing over running on command line
Instead of using npm test, this scheme now uses mocha for unit testing.
This has a profound advantage that, combined with c8, there is the
ability to run the code, assert values, *AND* get code coverage. The
code coverage has already proved helpful in finding some sead code in
the image reading code.