Improve arithmetic test coverage
This tests a bunch more arithmetic, including the transparent coversion
of SmallInt into their arbitrary precision counterparts (internally,
LargePositive and LargeNegative.)  It also increases code coverage from
94% to nearly 96%.

Most code that is not covered by tests at this point are impossible edge
cases (e.g., bytecode that should clearly be impossible). These aren't
easy to test, and are mostly left around as assertions due to bugs in
the interpreter (or, potentially, in the bytecode compiler in the
Smalltalk image itself).