Exercise error handling code (and GUI)
Currently, when an error is encountered in the system (such as sending a
message an object does not understand), the VM sends that object the
message #error: to handle that error. Currently, the implementation of
that message in Object opens the SmallWorld GUI with a text panel with
the error.

Since the system doesn't have a GUI yet, I implemented the GUI system
primitives as a spy object in the test class, so that there can be a
headless test that the GUI was popped up and displayed the correct error
message.

This scheme worked, but I may not stick with it, depending on the
eventual state of GUI handling.