Handle error tesing in a better way
The original definition of Object>>#error: was

error: str  | w t |
   t <- Pane textArea.
   t setText: (str + Char newline asString + Context current backtrace).
   w <- Window new title: 'Error'. w width: 300 height: 300.
   w setPane: (Pane north: nil center: t south: (Pane title: 'close' action: [w close])).
   w show.  Object halt

This is tricky because it relies on GUI functionality that neither
exissts in primitives nor makes sense in a unit test. Testing the code
for the GUI was out of scope for testing the VM.

This was simplified to

error: str
  ^ str + Char newline asString + Context current backtrace

And the resulting image was saved as testdata/image.noGui. This name is
misleading because the GUI wasn't completely removed, but the intention was to
have the test cases use this special prisitine image so it can exercise features
without needing the GUI primitives.

The end result was that the complicated spy could be unwound, so the test code
is simpler, yet the interpter VM coverage still rises from 96% to 98% line
coverage.