Bring test coverage up to 100%
In order to test some of the esoteric error cases in the VM, I
understand the bytecode well enough now to hand create incorrect
bytecode. This would only happen if the compiler generated bad bytecode,
but it's useful to test anyway (because one of the cases had an error
anyway.)

This means there is 100% test and branch coverage in the interpreter!