Remove unnecessary test images
Inspired by the scheme of patching the bytecode, all of the invalid
object tests were hand manipulated binaries. Rather than have several
hundred k of non-working images, now the correct image is patched in
place, which is more explicit about what is going on.

There is stil a test image (the one which redefines Object>>#error:) but
that's because binary patching that way is non-trivial.