Exercise running the UI (Adds action handlers)
The test cases now click on list items and buttons. This exercises new
functionality: callbacks from the UI, and actions. On the Smalltalk
side, Actions have blocks that should be run when the even fires. On
the interpreter side, the action blocks are turned into contexts that
are then executed.