Give the REPL a fake UI
Previously, if you used the REPL to perform any action that would cause
the GUI to open, it would crash on an unkown primitive. Now some hints
of the GUI are printed to the console.

This is useful for typing an error at the console (which previously
crashed the whole REPL). It is also helpful because some operations
required the GUI. Now, those operations will proceed.