Implement image saving primitive
The image saving primitive uses ImageWriter to create a new Uint8Array.
If the interpreter has an (optional) imageSaveCallback, it is called
with the new image. That way, image saving can be implemented in a node
or HTML specific way.

The repl implements this with a simple node call to writeFileSync.