Begin work on creating a no-GUI image
This is the first in seveeral steps of creating a GUI that does not
depend on the SmallWorld Java AWT/Swing inspired primitives. It is a
script of Smalltalk codde that removes or alters the necessary classes.
This can be run on the command line with

  ./repl.js < remove-gui.st

And produces an update in image-nogui.data.

This first version removes the hidden classes ListPane and TextPanel.
These were almost certainly incomplete in SmallWorld, as the Pane class
versions of these work better, and the code surrounding them (like nil
test) don't work properly in Java either.