Remove references to Semaphore
Semaphore depended on Java threads to halt execution of the interpeter.
In JavaScript, which is single threaded, this makes less sense.

Semaphores could probably be simulated in this JavaScript implementation
using continuations (pausing the state of the interpreter and wrapping
it in a context) but it probably makes more sense to use callbacks
directly because the style is more modern to today's eyes.