Implement sleep
1000 sleep sleeps for one second (1000 milliseconds). Although the
JavaScript system doesn't have threads and blocking, it does have
setTimeout.

The nice thing about the interpreter and the Smalltalk object model is
that the interpreter only needs an execution Context, and the Context
has sufficient information (it's byte codes, program counter, and parent
context for returning).

Because of this, it's easy to implement sleep as a continuation - put
all the state in context, and in the setTimeout, execute the saved
context. From the perspective of the Smalltalk virtual machine, nothing
happens until the bytecode continues, so this works as expected in
Smalltalk code.