Update repl multiline support
Previously, a ! began a multiline string, ending with another !. This
was good, but the core reason for having multi-line support was to
compile methods. The most subtle thing about this was that this
evaluates as Class compileMethod: 'body' where that argument needs to be
a literal string. This means that string literals (in the method body)
need to be written as ''a'' because of Smalltalk's escaping rules. This
was harder to read and easy to forget.

The new scheme is nice because it makes the text file more readable (you
say you want a METHOD and the class you want the method added to), and
it reads the body until ENDMETHOD. It also does the right thing to
escape the strings within.