Change how callbacks work
Previously, when JavaScript code needed to call back to SmallTalk code
(a button was clicked or a list item selected) the JavaScript code
generated a JavaScript number, and the interpreter converted that into a
SmallInt.

This new mechanism is more general purpose, as a block can be sent any
value (which requires that the middle layer do the integer -> SmallInt
converstion.

This doesn't matter much right now, but will pave the way for future
improvements.