Update the REPL to support meta-classes
Sometimes it is useful to edit the class methods, so an instance method
can be added with METHOD and a class method added with META.

There is some experimental code being checked in right now as well, for
a native HTML GUI in progress. This code adds a log primitive (which
will be removed).