Create an image completely removing GUI
With this version of the image, the final bits of the original
Java-centric GUI (Menu, Pane, Window) are not in the image. The
resulting image is truly headless, as it doesn't have primitives that
require a GUI.

Comparing this to Budd's original image, the new version is reduced by
nearly 50K, dropping over 29% in size from 172728 bytes to 122093 bytes.