Add a halt to error:
When the system encounters an error, it should call Object halt to stop
execution of that thread. Otherwise things keep running after the error.
That's fine for recoverable errors, but not for the generic case.

This also adds a logger to the standard nogui image, which is helpful
for testing.