Create a new, HTML-native UI
This introduces a new Smalltalk class called Element which represents an
HTML DOM element. It has methods for doing DOM manipulation, such as
creating elements, finding them in the document, addding and removing CSS
classes, and appending and removing from the DOM, and handling DOM
events with Smalltalk blocks.

The end result is a new UIHandler which does basically everything that
the old UI handler did, but with 10 primitives instead of 17, and a
great deal of flexibility that the old UI lacked.

This Element handler basically implements the same UI that the old one
did, but all within the classes (Pane, Menu, Window).

This also demonstrates the power of local variable based closures. The
list method creates an HTML-style ListBox with only a tiny amount of
code. This works by creating a click handler on list items that executes
on the local variable state.

This is not yet integrated with the Web interface, but doing so should
be relatively straightforward.