Rename browser
It should be noted that the class browser and class editor in the new
version are intentionally very close to the previous Java version, just
with different implementations of the factories that make widgets. This
isn't the ideal state, but proves that the GUI can be completely swapped
out with alternative implementations without much work.