Add a browser-basaed GUI test
If you run guitest.html in a browser, it runs an in-browser mocha test.
This is useful because the new GUI does direct DOM manipulation. This
new test performs DOM queries and manipulations to make sure that the
new UI performs. Because compiling and running exercises much of the VM,
it also serves for a fairly through test of the VM in a browser in an
automated way.