Create a new "SmallWorld" runner class
Previously, a considerable amount of logic was duplicated in the HTML
versions, the REPL, and the various tests. All of this logic related to
starting the interpreter and running methods.

This introduces a new class SmallWorld that does all of this at once,
and refactors everything to use that new class. That means that each
file now has far fewer dependencies, and is a lot smaller because the
random versions of runDoIt have been consolidated.

The new SmallWorld class takes as input an array buffer (a Uint8Array)
which is produced somehow (either from the fetch API in the browser, or
the fs primitives in node). It then provides the interpreter, a doIt,
and a printIt method.