Move image saving out of interpreter
The nice thing about having the SmallWorld runner class is that the
interpreter itself has no dependencies on the object file format, it is
given objects and operates on them. With this change, the interpreter no
longer depends on image reading or writing, that task being delegated to
SmallWorld.

SmallWorld can optionally get passed a callback funciton, and that
callback function receives a file name and a Uint8Array to be saved.
SmallWorld registers this with the interpreter, which calls it when the
save primitive is called.