Experiment with recompiling the whole image
The script in experimental/list.st was used to generate some stats, but
perhaps more imporantly, attempt to recompile every object in the image.

There were some tiny differences in the bytecode of a few methods
(probably owing to the way that temporary variables were calculated for
blocks) but otherwise were pretty similiar.

The interesting difference is that the resulting image grew in size,
from 122338 bytes in image-nogui.data to 127421 in nogui-experimental.
The cause for this 5083 byte size increase is unclear, and reading a
binary image is difficult.

So, before I re-use the new fully recompiled version, I want to
investiage why it didn't re-create the image identically.