Remove experimental image docs
The root cause for why the image sizes are different are due to unique
strings.

Looking at the original and new image, we wee the following instance
counts:

Class       Old     New
SmallInt    392     290
Array       460     497
String     1704    2022

With the exception of SmallInt, I think what happened here is that the
old image used to re-use identical objects, especially strings. If you
look at the strings in the old and new image, they are identical, but
the new image has duplicate copies of the strings.

So, I don't think the new image is corrupt, it just does less duplicate
removal than the original.