Provide a method to dump an image file textually
This adds a dumpToText method to ImageWriter which logs every object in
the image with it's class, its instance variables, and (for integers and
strings) their value.

It would be interesting to check in the resulting text file, but it's
close to 500K so it seems wasteful.

This is useful for debugging image and compiler optimizations.