Remove duplicate objects in image storage
When looking at the image data and compiled methods, I noticed that
similar objects kept being repeated. My investigation of the compiler
showed that the same literals got added again and again. A fix for that
issue is forthcoming, but it caused a deeper dive into the image.

Looking at the image, some interesting things emerged:

* The empty array (an array with size 0) appeared 108 times
* The same SmallInts appeared again and again, with only 445 distinct
  SmallInts in the image
* Because in part of the intense duplication of strings in object
  literals, 1553 strings appeared more than once in the image.

So, ImageWriter has been updated to remove duplication of these
objects. So, each time a string or a number is encountered, it is
added to a map. If an identical string or number is encountered, the
first version of the string is used instead (by mapping the old object
index to a different one).

For the empty array and for SmallInts, this is completely safe, since
the objects are immutable, so there can be as many references to a
shared SmallInt as desired.

Strictly speaking, this is not safe for Strings, because strings are
mutable. So, if the string "add" appeared in the image, and someone
modified that shared instance (e.g., s at: 1 put $d) then all of
the references will point to the string "ddd". That said, strings
are typically used in a way where mutability doesn't matter.

In fact, in some places, strings were *already* being re-used. Looking
carefully at some of the duplicated objects, many are message
selectors (e.g., at:put:). This isn't possible in the current
SmallWorld, but this does appear in Budd's original image (albeit
inconsistently). How this happened is unclear; perhaps the original
image was compiled with a different compiler that interned symbols.
Nevertheless, things were working properly with reused strings before,
and the new packed version still passes all the unit tests.

The savings of this approach are dramatic. Budd's original image
shrunk from
   172729 bytes and 4903 distinct objects
to
   134738 bytes and 2800 distinct objects.

Which makes the image nearly 22% smaller, with nearly 43% fewer
objects. The no GUI image is now less than 100K.

The result also reduces in-memory usage, especially because of the
redundancy with object literals, which appear many times.