Reinstate fully recompiled image
The problem with Object>>#hierarchy wasn't that the bytecode was bad.
Instead, it turned out that the original source in Class had syntax
errors that were not being caught by the compiler. It turns out that the
statement before a return statement was missing a period, which was an
error and producing bad code.

I inpsected all the code in the image and found a second example of this
(interesting enough, also in Class). After correcting this error, Object
hierarchy works, and with the unit tests added for Array and List, all
the other methods whose bytecode changed should also be tested. Thus,
the recompiled image (with bug fixes included) is now reinstated.