Fix various types of equality tests
For most non-numbers, = does what you expect, returning fals if the
things are not the same object (defined by Object>>#=). However, when
numeric types are compared against non-numbers, they used to return
errors, e.g., (3 = 'a') would raise an error because it attempted to
perform a SmallInt comparison to the string.

To fix this, a few of the missing comparison messages have been added to
Object, as any non-number compared to Object will be false.

The odd case is that LargeNegative comparisons rely on calling #negated
which doesn't make sense for objects. The simplest solution to this
problem was to have Object negated return false. This feels a bit weird
because 'foo' negated should really raise an error, not return null.