Better equality tests
Object negated felt weird, as it really should return an error. An
alternative, more Smalltalk-y solution is for all objectes to return
false to #negatable, but numbers return true. This, when combined with
an update to LargeNegative's equality test (to use the type guard) makes
it so that the selectors that objects respond to make sense.