Optimize the compiler
Previously, when the compiler generated the array of literals, each
literal seen in the source code was naively appended to the list of
literals (the literal array is part of the compiled method object).

For most small methods, this is fine, but large methods tend to use the
same literals again and again, and the literal array can get quite
large.

This leads to a modest size savings in the nogui image down from 99934
to 99309 bytes, but a larger savings in the GUI image, which dropped
nearly 1% in size from 108068 to 107123 bytes. Large methods like
Class>>#edit see large savings, with the literals array shrinking from
75 to 41 elements (because things like Element and button: are
referenced many times).