Add experimetnal canvas support
This code is not yet working, but begins support for drawing on an HTML
canvas. This is an interesting technical problem because there are so
many methods on the canvas API. The approach taken is to define a single
primitive (70) which contains an opcode and a variable number of
arguments. It's the responsibility of the canvas primitive handler to
pop the proper number of arguments off the stack.