Add Canvas support to the HTML UI
This exposes the HTML 2D Canvas API to SmallWorld. It adds about 8K to
the GUI image (including examples).

Canvas is interesting because there are so many methods for various
types of drawing. To avoid creating dozens of primitives, I created a
single new primitive (number 70) where the top of stack is an opcode
representing the canvas drawing command. The native handler then pops
the proper number of items off the stack for that particular call.