Update image format to allow floats
Previously, Float (real numbers) couldn't be stored in the image file
format. In practice, this meant that real numbers could not appear in
source code, because the Flaot would be added to the list of literals,
but then the literal could not be serialized.

This adds a new object type (3) for Float SmallJsObjects. The serialized
format of a float is a string (the decimal representation of the float,
as JavaScript stores it.) So it is saved by float.toString() and
restored with parseFloat(s).

I considered updating the version number in the image file format, as
this can now produce images that can't be read with the Java version.
However, I don't really intend to support that (as the primitives have
already diverged), and this new version is backwards compatible with the
last, so version 0 it remains.